+++ src/main/java/com/jug/config/ConfigurationManager.java
@@ -225,7 +225,15 @@ public class ConfigurationManager implements ITrackingConfiguration, IUnetProces
      */
     public boolean EXPORT_ORIENTED_BOUNDING_BOX_MEASUREMENT = true; /* set whether to perform the oriented bounding box measurement */
 
+    /**
+     * Sets if the crossing constraints should be used.
+     */
     public boolean USE_FEATURE_CROSSING_CONSTRAINTS = true;
+
+    /**
+     * Sets if the crossing constraints should be used.
+     */
+    public boolean USE_FEATURE_MIGRATION_COSTS = false;
 
     private int minTime = -1;
     private int maxTime = -1;
@@ -309,6 +317,8 @@ public class ConfigurationManager implements ITrackingConfiguration, IUnetProces
         INTENSITY_FIT_INITIAL_WIDTH = Double.parseDouble(props.getProperty("INTENSITY_FIT_INITIAL_WIDTH", Double.toString(INTENSITY_FIT_INITIAL_WIDTH)));
 
         USE_FEATURE_CROSSING_CONSTRAINTS = parseBooleanFromIntegerValue("USE_FEATURE_CROSSING_CONSTRAINTS", USE_FEATURE_CROSSING_CONSTRAINTS);
+
+        USE_FEATURE_MIGRATION_COSTS = parseBooleanFromIntegerValue("USE_FEATURE_MIGRATION_COSTS", USE_FEATURE_MIGRATION_COSTS);
 
         /* process feature flags */
         if (featureFlagDisableMaxCellDrop) {
@@ -450,6 +460,9 @@ public class ConfigurationManager implements ITrackingConfiguration, IUnetProces
             props.setProperty("SPINE_MEASUREMENT_POSITION_AVERAGING_MAXIMUM_WINDOWSIZE", Integer.toString(SPINE_MEASUREMENT_POSITION_AVERAGING_MAXIMUM_WINDOWSIZE));
 
             props.setProperty("SPINE_MEASUREMENT_MEDIALLINE_OFFSET_FROM_CONTOUR_ENDS", Double.toString(SPINE_MEASUREMENT_MEDIALLINE_OFFSET_FROM_CONTOUR_ENDS));
+
+            setBooleanAsIntegerValue(props, "USE_FEATURE_CROSSING_CONSTRAINTS", USE_FEATURE_CROSSING_CONSTRAINTS);
+            setBooleanAsIntegerValue(props, "USE_FEATURE_MIGRATION_COSTS", USE_FEATURE_MIGRATION_COSTS);
 
             props.store(out, "MotherMachine properties");
         } catch (final Exception e) {
@@ -597,5 +610,8 @@ public class ConfigurationManager implements ITrackingConfiguration, IUnetProces
         return new Version(getDatasetMomaVersionString());
     }
 
-    public boolean getCrossingConstraintFeatureFlag() { return USE_FEATURE_CROSSING_CONSTRAINTS; };
+    public boolean getCrossingConstraintFeatureFlag() { return USE_FEATURE_CROSSING_CONSTRAINTS; }
+
+    public boolean getMigrationCostFeatureFlag() { return USE_FEATURE_MIGRATION_COSTS; }
+
 }