+++ src/test/java/com/jug/feature/Feature__20221013_implement_non_crossing_constraint.java
@@ -1,4 +1,4 @@
-package com.jug;
+package com.jug.feature;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;